# apps/docs-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Claustrum Memory Core API",
    "version": "0.1.3",
    "description": "Auto-generated from memory-core route declarations and inline Zod validations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/v1/project-members": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "GET /v1/project-members",
        "operationId": "get_v1_project_members",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      },
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "POST /v1/project-members",
        "operationId": "post_v1_project_members",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "project_key": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "OWNER",
                      "MAINTAINER",
                      "WRITER",
                      "READER"
                    ]
                  }
                },
                "required": [
                  "workspace_key",
                  "project_key",
                  "email",
                  "role"
                ]
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{key}/members": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "GET /v1/projects/:key/members",
        "operationId": "get_v1_projects_key_members",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      },
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "POST /v1/projects/:key/members",
        "operationId": "post_v1_projects_key_members",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "OWNER",
                      "MAINTAINER",
                      "WRITER",
                      "READER"
                    ]
                  }
                },
                "required": [
                  "workspace_key",
                  "email",
                  "role"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/projects/{key}/members/{userId}": {
      "put": {
        "tags": [
          "Project"
        ],
        "summary": "PUT /v1/projects/:key/members/:userId",
        "operationId": "put_v1_projects_key_members_userId",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "OWNER",
                      "MAINTAINER",
                      "WRITER",
                      "READER"
                    ]
                  }
                },
                "required": [
                  "workspace_key",
                  "role"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "DELETE /v1/projects/:key/members/:userId",
        "operationId": "delete_v1_projects_key_members_userId",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      }
    },
...
    "/v1/users": {
      "get": {
        "tags": [
          "Access"
        ],
        "summary": "GET /v1/users",
        "operationId": "get_v1_users",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      },
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "POST /v1/users",
        "operationId": "post_v1_users",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{key}/api-keys": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "GET /v1/workspaces/:key/api-keys",
        "operationId": "get_v1_workspaces_key_api_keys",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "POST /v1/workspaces/:key/api-keys",
        "operationId": "post_v1_workspaces_key_api_keys",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "label": {
                    "type": "string"
                  },
                  "device_label": {
                    "type": "string"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "user_id",
                  "device_label"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{key}/api-keys/{apiKeyId}": {
      "delete": {
        "tags": [
          "Workspace"
        ],
        "summary": "DELETE /v1/workspaces/:key/api-keys/:apiKeyId",
        "operationId": "delete_v1_workspaces_key_api_keys_apiKeyId",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKeyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      }
    },
...
    "/v1/workspaces/{key}/members": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "GET /v1/workspaces/:key/members",
        "operationId": "get_v1_workspaces_key_members",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "POST /v1/workspaces/:key/members",
        "operationId": "post_v1_workspaces_key_members",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "OWNER",
                      "ADMIN",
                      "MEMBER"
                    ]
                  }
                },
                "required": [
                  "email",
                  "role"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{key}/members/{userId}": {
      "put": {
        "tags": [
          "Workspace"
        ],
        "summary": "PUT /v1/workspaces/:key/members/:userId",
        "operationId": "put_v1_workspaces_key_members_userId",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "OWNER",
                      "ADMIN",
                      "MEMBER"
                    ]
                  }
                },
                "required": [
                  "role"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workspace"
        ],
        "summary": "DELETE /v1/workspaces/:key/members/:userId",
        "operationId": "delete_v1_workspaces_key_members_userId",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-primary-routes-access.ts"
      }
    },
...
  },
...
}